Moved participants_file function to bids.py
